# Faster hmget when fetching a lot of fields